requirements.txt
python-telegram-bot==13.15
pytz
requests
tradingview_ta==3.2.7